backend_v2/pyproject.toml
[project]
name = "spm-backend-v2"
version = "2.0.0"
description = "SPM Backend v2.0 - API REST limpia"
requires-python = ">=3.11"
dependencies = [
    "flask==3.1.2",
    "flask-cors==6.0.1",
    "pyjwt==2.10.1",
    "python-dotenv==1.0.1",
    "pydantic==2.10.5",
    "pydantic-settings==2.7.1",
    "sqlalchemy==2.0.44",
    "itsdangerous==2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.0",
    "pytest-cov==6.0.0",
    "black==25.1.0",
    "flake8==7.1.2",
    "mypy==1.15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=backend_v2",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "flask_cors.*"
ignore_missing_imports = true
